+++ Changelog.md
@@ -45,6 +45,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
    - "row" always points to the "data" context object for that instance (typically row in DataGrid)
    - "data" always refers to the global data of the submission.
 
+## [UNRELEASED]
+### Changed
+ - Default value for calculated logic.
+
 ## 3.0.0-rc.8
 ### Fixed
  - Problem with the Select component where it would remove selected items while searching.